requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
Flask-Mail==0.9.1
PyMySQL==1.1.1
Pillow==10.4.0
reportlab==4.2.2
python-dotenv==1.0.1
Werkzeug==3.0.3
WTForms==3.1.2
qrcode==7.4.2
cryptography==42.0.8
greenlet==3.0.3
# eventlet et gunicorn sont utilisés uniquement dans Docker
# Décommentez si vous déployez sur Linux/Docker :
# eventlet==0.35.2
# gunicorn==22.0.0
# redis==5.0.8
